refactor(marketing): shared route-group layout for all marketing pages
