+++ src/app/[locale]/(unauth)/(marketing)/about/about-content.ts
