+++ src/app/[locale]/(unauth)/(marketing)/layout.tsx
@@ -3,21 +3,21 @@ import { MarketingFooter } from '@/components/marketing/footer';
 import { MarketingNavbar } from '@/components/marketing/navbar';
 
 /**
- * Demo route group for the opt-in marketing shell.
+ * Shared chrome for every marketing page (`/`, `/about`, `/blog`, `/changelog`,
+ * `/terms`, `/privacy`).
  *
- * Wraps children in the brand-config-driven `MarketingNavbar` + `MarketingFooter`
- * so a fork can SEE the shell rendered. This is ADDITIVE and does not affect the
- * existing `(unauth)` pages, which keep using `src/templates/Navbar`/`Footer`.
+ * This route group is the canonical marketing shell: it wraps all its children
+ * in the brand-config-driven `MarketingNavbar` + `MarketingFooter` so the header
+ * and footer are applied in exactly ONE place instead of per-page. Content pages
+ * inside the group render page content only — no site chrome of their own.
  *
  * The `AuthDialogProvider` is scoped HERE (not the parent `(unauth)/layout.tsx`)
  * because only this subtree consumes the overlay dialog. It must wrap BOTH the
  * `MarketingNavbar` (whose CTAs call `openSignIn`/`openSignUp`) AND the page
  * `children` (which mount `AuthDialogAutoOpener`) so both stay descendants of
  * the provider — otherwise the dialog never opens.
- *
- * A fork opts in by moving (or pointing) its marketing pages at this layout.
  */
-export default function MarketingV2Layout({
+export default function MarketingLayout({
   children,
 }: {
   children: React.ReactNode;
